--- a/src/java.base/unix/native/libjli/java_md.c
+++ b/src/java.base/unix/native/libjli/java_md.c
@@ -300,6 +300,19 @@ CreateExecutionEnvironment(int *pargc, char ***pargv,
     char * jvmtype = NULL;
     char **argv = *pargv;
 
+#ifdef AIX
+      const char *mallocOptionsName = "MALLOCOPTIONS";
+      const char *mallocOptionsValue = "multiheap,considersize";
+      if (setenv(mallocOptionsName, mallocOptionsValue, 0) != 0) {
+          fprintf(stderr, "setenv('MALLOCOPTIONS=multiheap,considersize') failed: performance may be affected\n");
+      }
+      const char * ldrCntrlName = "LDR_CNTRL";
+      const char *ldrCntrlValue = "TEXTPSIZE=64K@DATAPSIZE=64K@STACKPSIZE=64K";
+      if (setenv(ldrCntrlName, ldrCntrlValue, 0) != 0) {
+           fprintf(stderr, "setenv('LDR_CNTRL=TEXTPSIZE=64K@DATAPSIZE=64K@STACKPSIZE=64K') failed: performance may be affected\n");
+      }
+#endif
+
 #ifdef SETENV_REQUIRED
     jboolean mustsetenv = JNI_FALSE;
     char *runpath = NULL; /* existing effective LD_LIBRARY_PATH setting */