--- a/test/jdk/java/lang/ProcessBuilder/Basic.java
+++ b/test/jdk/java/lang/ProcessBuilder/Basic.java
@@ -799,11 +799,15 @@ public class Basic {
     }
 
     /* Only used for AIX --
-     * AIX adds the variable AIXTHREAD_GUARDPAGES=0 to the environment.
-     * Remove it from the list of env variables
+     * AIX adds the variables AIXTHREAD_GUARDPAGES=0,
+     * LDR_CNTRL=TEXTPSIZE=64K@DATAPSIZE=64K@STACKPSIZE=64K, and
+     * MALLOCOPTIONS=multiheap,considersize, to the environment.
+     * Remove them from the list of env variables
      */
     private static String removeAixExpectedVars(String vars) {
-        return vars.replace("AIXTHREAD_GUARDPAGES=0,", "");
+        return vars.replace("AIXTHREAD_GUARDPAGES=0,", "")
+                   .replace("LDR_CNTRL=TEXTPSIZE=64K@DATAPSIZE=64K@STACKPSIZE=64K,","")
+                   .replace("MALLOCOPTIONS=multiheap,considersize,","");
     }
 
     private static String sortByLinesWindowsly(String text) {