@AlanBateman 2024-02-19
Question: Is the java launcher the right place for this? I'm wondering about other launcher that uses JNI Invocation API to create the VM.