@dholmes-ora 2024-02-20
I concur with Alan: is this the right way to fix this? And is it really the case that every single Java application on AIX wants this setting?